create a snippet for testing, change test name, improve EMF handling
