+++ src/test/resources/snippets/procedure_text.xml
@@ -1,0 +1,1 @@
+<?xml version="1.0" encoding="UTF-8"?><?xml-stylesheet type="text/xsl" href="cda.xsl"?><ClinicalDocument xmlns="urn:hl7-org:v3" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:sdtc="urn:hl7-org:sdtc" xsi:schemaLocation="urn:hl7-org:v3 ../../../CDA%20R2/cda-schemas-and-samples/infrastructure/cda/CDA.xsd" classCode="DOCCLIN" moodCode="EVN"><realmCode code="US"/><typeId root="2.16.840.1.113883.1.3" extension="POCD_HD000040"/><templateId root="2.16.840.1.113883.10.20.22.1.1" extension="2015-08-01"/><templateId root="2.16.840.1.113883.10.20.22.1.2" extension="2015-08-01"/><id root="2.16.840.1.113883.3.552.1.3.100.1.13.1.999362" extension="4950042"/><code code="34133-9" codeSystem="2.16.840.1.113883.6.1" displayName="Summarization of Episode Note"/><title>Continuity of Care Document</title><effectiveTime value="20180720171812.204-0400"/><confidentialityCode code="N" displayName="Normal" codeSystem="2.16.840.1.113883.5.25"/><languageCode code="en-US"/><component typeCode="COMP"><structuredBody classCode="DOCBODY" moodCode="EVN"><component typeCode="COMP" contextConductionInd="true"><section classCode="DOCSECT" moodCode="EVN"><templateId root="2.16.840.1.113883.10.20.22.2.7.1" extension="2014-06-09"/><templateId root="2.16.840.1.113883.10.20.22.2.7.1"/><code code="47519-4" codeSystem="2.16.840.1.113883.6.1" codeSystemName="LOINC" displayName="History of Procedures"/><title>Procedures</title><text><table cellspacing="1" cellpadding="0" width="95%"><colgroup><col width="45%"/><col width="10%"/><col width="20%"/><col width="15%"/><col width="10%"/></colgroup><thead><tr><th>Procedure</th><th>Date</th><th>Related Diagnosis</th><th>Body Site</th><th>Status</th></tr></thead><tbody><tr ID="ITEM_PROCEDURE018"><td><content ID="PROCEDURE018">APPENDECTOMY LAPAROSCOPIC<sup>1</sup></content></td><td>3/26/18</td><td> </td><td> </td><td>Completed</td></tr><tr ID="ITEM_PROCEDURE034"><td><content ID="PROCEDURE034">REMOVAL IMPLANTED DEVICES FROM BONE<sup>2</sup></content></td><td>3/26/18</td><td> </td><td> </td><td>Completed</td></tr><tr ID="ITEM_PROCEDURE049"><td><content ID="PROCEDURE049">REPAIR ELBOW<sup>3</sup></content></td><td>3/26/18</td><td> </td><td> </td><td>Completed</td></tr></tbody></table><paragraph><sup>1</sup><content ID="PROCCMT027">auto-populated from documented surgical case</content></paragraph><paragraph><sup>2</sup><content ID="PROCCMT087">auto-populated from documented surgical case</content></paragraph><paragraph><sup>3</sup><content ID="PROCCMT071">auto-populated from documented surgical case</content></paragraph></text><entry typeCode="DRIV" contextConductionInd="true"><procedure classCode="PROC" moodCode="EVN"><templateId root="2.16.840.1.113883.10.20.22.4.14" extension="2014-06-09"/><templateId root="2.16.840.1.113883.10.20.22.4.14"/><id root="77baeec3-124e-4348-bcec-fbe2fd25e7ef"/><code nullFlavor="UNK"><originalText><reference value="#PROCEDURE018"/></originalText></code><statusCode code="completed"/><effectiveTime value="20180326091500.000-0500"/><entryRelationship typeCode="SUBJ" inversionInd="true" contextConductionInd="true"><act classCode="ACT" moodCode="EVN"><templateId root="2.16.840.1.113883.10.20.22.4.64"/><code code="48767-8" displayName="Annotation comment" codeSystem="2.16.840.1.113883.6.1" codeSystemName="LOINC"/><text><reference value="#PROCCMT027"/></text></act></entryRelationship></procedure></entry><entry typeCode="DRIV" contextConductionInd="true"><procedure classCode="PROC" moodCode="EVN"><templateId root="2.16.840.1.113883.10.20.22.4.14" extension="2014-06-09"/><templateId root="2.16.840.1.113883.10.20.22.4.14"/><id root="4372357a-30bb-48d9-a612-8f459ae8c00c"/><code nullFlavor="UNK"><originalText><reference value="#PROCEDURE034"/></originalText></code><statusCode code="completed"/><effectiveTime value="20180326104500.000-0500"/><entryRelationship typeCode="SUBJ" inversionInd="true" contextConductionInd="true"><act classCode="ACT" moodCode="EVN"><templateId root="2.16.840.1.113883.10.20.22.4.64"/><code code="48767-8" displayName="Annotation comment" codeSystem="2.16.840.1.113883.6.1" codeSystemName="LOINC"/><text><reference value="#PROCCMT087"/></text></act></entryRelationship></procedure></entry><entry typeCode="DRIV" contextConductionInd="true"><procedure classCode="PROC" moodCode="EVN"><templateId root="2.16.840.1.113883.10.20.22.4.14" extension="2014-06-09"/><templateId root="2.16.840.1.113883.10.20.22.4.14"/><id root="e39bdae7-7b35-4ceb-88b7-8891414c3bc2"/><code nullFlavor="UNK"><originalText><reference value="#PROCEDURE049"/></originalText></code><statusCode code="completed"/><effectiveTime value="20180326121000.000-0500"/><entryRelationship typeCode="SUBJ" inversionInd="true" contextConductionInd="true"><act classCode="ACT" moodCode="EVN"><templateId root="2.16.840.1.113883.10.20.22.4.64"/><code code="48767-8" displayName="Annotation comment" codeSystem="2.16.840.1.113883.6.1" codeSystemName="LOINC"/><text><reference value="#PROCCMT071"/></text></act></entryRelationship></procedure></entry></section></component></structuredBody></component></ClinicalDocument>
